backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
bcrypt>=4.0.0
google-genai>=1.0.0
openai>=1.30.0
chromadb>=0.5.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
pypdf>=5.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.5
pydantic>=2.10.0
pydantic-settings>=2.6.0
sse-starlette>=2.1.0
aiofiles>=24.1.0
unstructured[docx]>=0.16.0
